server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1 
openai==1.12.0
httpx>=0.24.1,<0.28
python-dotenv==1.0.1